ship Vidalia instead of TorK as Tor controller
TorK was designed for KDE 3 and not GNOME, and Vidalia is officially
recommended by the Tor project, we replace the former by the later.

As Vidalia is not well integrated to Debian, we use a custom package
(source attached to the bug report) that:
 - will never start an instance of Tor by itself,
 - lacks some UI elements useless in the context of amnesia.

There is no KDE based applications anymore, so `kde-i18-*` are removed from
packages lists.