Build rt2860 modules from Squeeze.
This supports the RT2860 wireless adapter, found particularly in the ASUS EeePC
model 901 and above.

Converted 51-rt2570-modules hook to a generic module-assistant one, fixing one
bug on the way.

Added non-free category to the Squeeze sources, as the rt2860-source package
is there.

Force installing debhelper, as rt2860-source effectively depends on it to build.