Install version 2.0~a21-1 of live-build and live-helper in the image.
Our "normal" pinning installs the versions found in Squeeze.
2.0~a29-1 has migrated to Squeeze.
It changes quite a lot of things (e.g. rename LH_* variables to LB_*)
which we cannot afford dealing with at this stage of the release process.
Indeed upgrading to 2.0~a29-1 would probably imply:

- adapting various bits of our build process (hooks and so on)
- re-doing the testing that has already done on T(A)ILS 0.6~rc1.

Temporary solution: include version 2.0~a21-1 of the live-build and live-helper
packages in our source tree and setup APT pinning so that they are preferred to
the version in Squeeze.