No need to use absolute path here.
Using absolute paths may be needed for code that should run after the boot
medium has been removed. Only such code in this script is the do_stop function
=> no need to use absolute paths elsewhere.

On the other hand, the initscript that calls us sets a pretty scarse $PATH so we
need to add /usr/bin to it.