Revert "Bring Broadcom STA modules back."
This reverts commit d3ee66bb5a04c27cff2f24eda011267022516d05.

I was confused, which is not surprising considering the shape of the Broadcom
drivers landscape (http://wiki.debian.org/bcm43xx#Drivers).

The broadcom-sta (wl.ko) driver indeed brings support for some recent WiFi
adapters, but it doesn't content itself with conflicting with the module that
supports the same hardware (brcm80211, brcmsmac): it also conflicts with the ssb
module that is itself needed by a few other modules such as b43, b44. In short,
it seems like wl.ko breaks more hardware support than it brings.