Set extensions.torbutton.prompted_language to true.
Else, Iceweasel asks very soon if the Torbutton "spoof English" setting should
be enabled, when Tails is started in !English language. We've enabled this
feature for quite some time, with reasons, and don't want 1. to nag the user
with such choices every time s/he starts Tails 2. to help users to shoot
themselves in the foot.