VirtualBox: add backported guest tools built against the ABI of X.Org backports.
The new version is 4.1.6-dfsg-2~bpo60+1.

The .deb's are not put in config/chroot_local-packages/; if they were,
live-build would try to install them early, which would fail as they depend on
a backported X.Org that we install in a local hook.

Remove the pinning too: else, it favoured packages from squeeze-backports in
a way that had our custom backport removed at the end of the build.