Set time to middle of [valid-after, fresh-until] from consensus.
According to dir-spec.txt a new consensus is published at fresh-until,
which is valid-after + 1 hour. Since we always fetch a new consensus
on boot we can narrow the time points we set the time to to the middle
of [valid-after, fresh-until], and since fresh-until = valid-after + 1
hour, that means we set the time to valid-after + 30 minutes.

The benefit of this is that if htpdate fails then the user still gets
a time that is at most 30 minutes incorrect. This, incidentally, will
prevent the known problem with hidden services refusing connections.