vagrant: Start a Rakefile to drive builds
Build based on Vagrant will be drived using `rake`. This has the
definite advantage that we can easily isolate Vagrant related files
in the `vagrant` subdirectory while providing an easy access to build
functions.

This will also allow easier local customizations to Vagrant itself,
as we will be able to mangle stuff in Rake before handing commands to Vagrant.

The Rakefile currently only wraps a few Vagrant commands, it will be extended
later on.