vagrant: Add task to create our squeeze basebox
As we use Veewee to automate the creation of the basebox, we are forced to back
up on the "do not execute Vagrant manually" stance. This is due to the fact that
Veewee is quite hackish and itself calls `vagrant` directly. So we move our
monkey-patching from the Rakefile to the Vagrantfile itself.

In order to support an HTTP proxy, the `preseed.cfg` file that Veewee need is
created beforehands in another Rake task. Proxies bound to localhost will
result in a failure and an explicit error messages.

Note: it is unfortunately not possible to use `definitions.rb` to add
`mirror/http/proxy` to the installer command-line. As the defined proxy is
set by network-preesed, it will prevent the installer from retrieving
the preseed file provided by Veewee on a local interface.