vagrant: Add build task
We export some environment variables to the build script inside the virtual
machine. One of them is http_proxy, so we first validate if it is not set
to use a proxy on localhost.

In order to properly get output from Vagrant, we handle its replication on
stderr/stdout manually.