Leave halt and reboot scripts configured
Given we provide a proper Required-Stop header, we can leave the halt and
reboot initscript in place. They will never be called as long as `tails-kexec`
succeed. If it don't, then it still provides a better user experience to
shutdown or reboot the machine than to leave the computer hanged.