Patch initscripts headers instead of fiddling with update-rc.d
Calling update-rc.d manually is quite error prone: for example, 'ttdnsd' is
actually required to be started after 'tor', but this was not reflected in the
number sequence.

Instead, we now patch the initscripts and rely on insserv to figure out the
correct ordering for the startup and shutdown sequence.

The patch to kexec-load is adjusted for consistency.