Fix tordate vs. Tor 0.2.3.x.
Since Tor 0.2.3.x Tor doesn't download a consensus for clocks that are
more than 30 days in the past or 2 days in the future (see commits
f4c1fa2 and 87622e4 in Tor's git repo). For such clock skews we set
the time to the Tor authority's cert's valid-after date to ensure that
a consensus can be downloaded.