Set the iceweasel torbrowser.version user_pref to "Tails".
According to my initial analysis, this is needed so that torbutton knows we're
running a torbrowser-patched browser, and hence switches to the appropriate
operation mode. From reading the torbutton code, setting this to a string that
is true when testing getCharPref in boolean context should be enough, since we
won't use the actual version number -- we have disabled version checking.