Fix Monkeysphere incompatibilities with Torbrowser patches
When Torbrowser patches are applied and the `socks_remote_dns` set to true,
Iceweasel will not allow any name resolution that does not go through a proxy.

This means that the default value of MONKEYSPHERE_VALIDATION_AGENT_SOCKET of
http://localhost:6136/ will not work ('localhost' needs to be resolved).
So we add a patch to make MVSA bind the IPv4 instead of 'localhost'.

Now that Torbutton detects Torbrowser patches, it enforces proxying more
strictly. The Monkeysphere extension then needs a new passthrough in
FoxyProxy rules to properly get a direct connection to the agent.