Remove config/chroot_local-packages/ from .gitignore.
The documented way for "external" contributors to add custom packages is to put
them in chroot_local-packages, and once we pull we import any such package into
our APT repo and rewrite the history appropriately.

Also, the ability to add packages in there and not see them in "git status"
makes it very easy to build tainted ISO images with non-standard packages,
which makes some of us fear can lead to hard to debug situations.