Disable IPv6 on all network interfaces.
This is a workaround for the IPv6 link-local multicast leak that was recently
discovered. Tails has no local service that listens on IPv6, so there should be
no regression, hopefully, unless one wants to play with OnionCat and VoIP,
but those of us should know how to workaround this anyway.