On boot, unblock bluetooth, wifi, wwan and wimax; block every other type of wireless device.
The ticket says to leave Bluetooth alone, but if we do that and it's disabled by
default by the firmware or whatever, then it would have no chance to get
connected devices, and then the design we decided for
todo/protect_against_external_bus_memory_forensics could not be implemented for
this protocol. So, we unblock it.