Don't install xul-ext-monkeysphere anymore.
As decided on todo/monkeysphere_keyserver_policy, because: the implication of
the current settings are not well understood, Monkeysphere is little used in
Tails, and we're not sure anymore it would be our first bet for the web browser
profile with no CA.

Let's keep the various configuration bits (e.g. FoxyProxy, patching MSVA),
though, so that advanced users who are used to have Monkeysphere in Tails just
have to install the package.