Import i2p's router.config from the package.
We're going to modify it quite a bit, for various reasons, at the same location,
so using a set of chroot_local-patches is not practical. Given this file, as
shipped by the package, contains only one line (that is specific to packaged
versions), it looks almost reasonable (though ugly and not totally future-proof)
to fork it.