Switch host's remote shell interface from pseudo tty to TCP.
This way avoid libvirt permission issues for the created tty:s (and
unix domain sockets if using that approach). Also, note that the
remote shell still uses a serial port on the guest, so this is not a
revert back to the approach requiring firewall exceptions inside the
Tails guest.