Capture Timeout::Error properly in try_for().
We want to ignore any exception of the yielded code block, but if the
Timeout::Error from the timeout wrapping the loop is thrown when the
code block is being executed that may lead to an infinite loop (if the
code block never returns true). Hence we need to handle that exception
differently. This assumes that the code block cannot throw
Timeout::Error.