Categorize IP vs IPv6 packets more reliably.
o.instance_of?(c) tests if o has the exact class c, but o.kind_of?(c)
also tests if o's is an instance of some subclass of c.

Also, the comment removed by this commit was wrong; TCPPacket *is* a
subclass if IPPacket, and so are {UDP,ICMP}Packet.