Make Tails shutdown steps less confusing.
The `I safely shutdown Tails` and `I shutdown Tails` were both
confusing names considering what they do. Merging the former into the
step when we wait for the memory wipe to finish, and renaming the
latter to say "completley" as in "we also wait until the computer
shuts down completely" hopefully make this clearer.