Move configuration of global vars into config.rb.
This way they're all set before any hooks are run. This is important
because we can't know in which order the hooks are run, so if we want
to compare some of these global variables that are set in two
different hooks, which we want to do in a commit to come, we're in
trouble.