Split VM networking into its own class, VMNet.
This split makes the setup of all the libvirt stuff on
feature/scenario start and end cleaner. Now we create one new VMNet
per @product-tagged feature, and let it live throughout that feature,
after which it is killed (same goes for the libvirt connection by
necessity). This has some very nice reliablity and performance
advantages compared to creating a VMNet per scenario (which would be
equivalent to what we've been doing until now):

We create a new VM instance (i.e. libvirt domain) per scenario
(possibly restored from a snapshot), but that's not necessary for the
VMNet instance; it could just as well survive for the whole
feature. One good reason for the latter is that restoring a VM and
have it use another VMNet instance than it was saved from seems to
confuse libvirt in such a way that re-establishing the network takes
extra time compared to restoring it to the same VMNet
instance. Another good reason is that network sniffing becomes less
decoupled from the VM instance, and can happen at any time.

(FWIW, this is a remnant from the good ol' unmerged test/reorg
branch.)