Fix buggy "persistence is not enabled" step.
It grep:ed for the wrong device (DM mappings are involved) so it would
pass even if persistence was disabled. Now this check uses the
persistence state file created by Tails Greeter in a way similar to
how tails-persistence-setup determines whether persistence is in use.
Also, this step's positive version now use the same approach, and both
steps were renamed to "is enabled/disabled" for greater consistency.

The old code for this step, which checks whether all persistent
directories have been successfully mounted, is put into a step of its
own, since it's still something we should test.