Allow the tails-persistence-setup user to modify internal storage devices through udisks.
Allow device change and mount. This is done for *internal* devices only, as we
now set the boot medium as 'system internal' for udisks, and t-p-s only needs to
touch the Tails boot medium in the current state of things.

LUKS unlock is allowed too, for both internal and external devices,
due to not-so-fine-grained available access rules.