Remove origin/ prefix in the branch name passed by Jenkins.
The Jenkins SCM plugin syntax for branch specifiers is REPO/BRANCH. It has some
helpful shortcuts, but in some cases (e.g. branches with a '/' in their name) we
really need to specify origin/BRANCH.

When origin/BRANCH is checked out, our custom APT source lists builder is
confused and falls back to the `devel' APT suite, instead of using the branch's
suite. So, we drop the `origin/' prefix from $REV before it's checked out.