Fix Torbrowser green onion checking, while preventing Torbutton from asking users to "upgrade TBB".
First, unset TOR_CONTROL_PORT and TOR_CONTROL_HOST (non-working since we are
forbidding access to the control port for the amnesia user), so that the "Are we
connected through Tor" check is done to check.torproject.org, instead of failing
to do it via the local control port.

Second, enable extensions.torbutton.versioncheck_enabled, set
extensions.torbutton.updateNeeded to false, and set
extensions.torbutton.lastUpdateCheck far away in the future. The former is
required so that the "Are we connected to Tor" check is tried at all, and the
two others are needed to actually block Torbutton's version checking, and
pretend no update was needed last time it checked.