Have live-boot's get_custom_mounts really read the new persistence.conf.
If we pass it the same $custom_mounts destination file on second invocation
(post-checking and potentially disabling of config files) as last time
(pre-checking), then we get the old, untrusted content, and then even the
persistence settings we have disabled are activated.

Note that we were doing the same mistake in the 0.21 automatic upgrade code,
but it was harmless back then: persistence_list="persistence.conf" prevented
live-boot from reading settings from the old live-persistence.conf in the first
place, when get_custom_mounts was run for the first time.