Get back to installing Linux 3.10-1.
This reverts commits 945796fea288a4abf074400b03ed02fe740f1fb5
and 6bebc91cef3d008683b4b5013b8490e374f3c9bd.

The APT pinning for linux*3.10* is actually useless, as these packages are not
in Debian testing anymore and will be covered by the general pinning rules that
apply to our own APT repository. Let's keep these lines around so that we don't
forget to adapt them once we move forward to a newer kernel again.