Use a Pidgin wrapper that initializes the GStreamer registry first.
Otherwise, if no GStreamer app was started before Pidgin, it segfaults the first
time, and works fine the second.

This workarounds Tails#5672, that is Debian#578476, which will apparently be
fixed once Debian ships a newer GStreamer, that is in Jessie.

While we're at it, move into this this wrapper the
GNOME_DESKTOP_SESSION_ID mangling.