Repeatedly press Enter until the GNOME Run dialog vanishes.
I have seen some weird race condition cases where the GNOME Run dialog has the
command to run ("gnome-terminal" in my case) correctly typed in, the text cursor
at the end, and the command step was not run, so the step fails with
"FindFailed: can not find GnomeTerminalWindow.png on the screen."

Issue with completion? Too slow or too fast test system? Bug somewhere in the
Sikuli / Ruby stack? Anyway, let's just add guards against this case and be done
with it.