Workaround race conditions in the encryption feature.
First, the pinentry window is not automatically focused in the "Symmetric
encryption and decryption using gpgApplet" feature, for some reason, while it is
focused in the others. So, we click on it to give it focus.

Second, on a fast (bare metal) test setup, when one immediately types the
passphrase, the first character is lost. So, we had a little sleep time. I have
not found any *visible* change during this time, so I can think of no other way
we detect when pinentry is ready to receive our input.