Add a startup script for the control port filter
The control port filter will be started as the vidalia user,
as the permissions it requires matches well with that of
the vidalia user.

It is always started, and does never need to be restarted.
If Tor is not running when a newnym request is made, it will
just return back an error message.