Remove the macchanger_helper() wrapper.
After finding a couple of bugs and issues in it I think we should just
KISS and not try to be so smart. This script is supposed to be run
immediately after the interface is plugged by udev, so the interface
should always be down. If it's up that's actually worrisome...