Merge three steps together and reorder them.
We merge "Tor has built a circuit", "the time has synced" and "Tor is ready"
into a single step called "Tor is ready". This simplifies scenarios quite a bit,
and allows us to always do the checks in the right order, without each scenario
trying its own version.

Speaking of the right order: we wait for the "Tor is ready" notification first.
it disappears after a timeout, so if we wait for other events first, we
sometimes cannot find the notification picture on screen later on.