Do not purge bind9-host's dependencies.
We did that in Squeeze for unclear reasons (commit 129b6cf). My guess
is that it has to do with the bind9 Squeeze backport, and that some
other package pulled them, resulting in multiple versions of these
deps being installed. These particular versions are in Wheezy so now
they shouldn't be removed any more.