Fix buggy memory checks for RAM building.
We have to take into account which state the Vagrant VM is in for
determining *where* we check if enough memory is available for a RAM
build. If it's off, we check the host; if it's on we check the
VM. Previously we always checked the host, which doesn't make sense
when the VM is already started.