Create tails-upgrade-frontend's trusted GnuPG homedir with stricter permissions, then make it looser. (Closes: #7037)
The goal here is merely to avoid the boot-time warning reported in #7037:  this
directory and its content need to be read-only for the tails-upgrade-frontend
user, so we're creating it as root. Let's create the directory with stricter
permissions, so that GnuPG does not complain when importing the key. And then,
let's chmod it to make it world-readable anyway.